libs/k3scompose/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "k3scompose"
version = "0.1.0"
description = "K3s Platform CLI for Docker Compose projects"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [{ name = "K3s Platform" }]
dependencies = [
    "pyyaml>=6.0",
    "jsonschema>=4.20",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
k3scompose = "k3scompose.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["k3scompose"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
